mini-rag-ui/requirements.txt
sentence-transformers
faiss-cpu
openai
numpy
PyPDF2
python-docx
pandas
openpyxl